+++ src/mocks/handlers/event.ts
@@ -1,22 +1,8 @@
+import { HttpResponse, http } from 'msw';
+import { environments } from '~/config/environments';
 import { ReadEvent } from '~/types/event/event';
 import { EventList } from '~/types/event/eventList';
 import { PageData } from '~/types/pageData';
-
-export const readEventMock: ReadEvent = {
-  id: 123,
-  mentorId: 123,
-  title: '웹 개발자 이력서 5일 간 첨삭',
-  content: '프론트, 백 관계 없이 웹 개발을 희망하시는 분들 이력서 피드백 드릴게요',
-  maximumCount: 3,
-  currentApplicantCount: 1,
-  status: 'OPEN',
-  positions: ['FRONT', 'BACK'],
-  timeInfo: {
-    openDateTime: '2024-04-24T08:00:00.000Z',
-    closeDateTime: '2024-04-30T23:59:59.999Z',
-    endDate: '2024-05-01T23:59:59.999Z',
-  },
-};
 
 const pageDataMock: PageData = {
   first: true,
@@ -32,7 +18,23 @@ const pageDataMock: PageData = {
   totalElements: 100,
 };
 
-export const eventListMock: EventList = {
+const readEventMock: ReadEvent = {
+  id: 123,
+  mentorId: 123,
+  title: '웹 개발자 이력서 5일 간 첨삭',
+  content: '프론트, 백 관계 없이 웹 개발을 희망하시는 분들 이력서 피드백 드릴게요',
+  maximumCount: 3,
+  currentApplicantCount: 1,
+  status: 'OPEN',
+  positions: ['FRONT', 'BACK'],
+  timeInfo: {
+    openDateTime: '2024-04-24T08:00:00.000Z',
+    closeDateTime: '2024-04-30T23:59:59.999Z',
+    endDate: '2024-05-01T23:59:59.999Z',
+  },
+};
+
+const eventListMock: EventList = {
   events: [
     {
       info: readEventMock,
@@ -53,3 +55,12 @@ export const eventListMock: EventList = {
   ],
   pageData: pageDataMock,
 };
+
+export const handlers = [
+  http.get(`${environments.baseUrlEnv()}/v1/events`, () => {
+    return HttpResponse.json(eventListMock);
+  }),
+  http.get(`${environments.baseUrlEnv()}/v1/events/:id`, () => {
+    return HttpResponse.json(readEventMock);
+  }),
+];
