+++ src/mocks/handlers/index.ts
@@ -1,0 +1,5 @@
+import { handlers as eventHandlers } from './event';
+import { handlers as resumeHandlers } from './resume';
+import { handlers as userHandlers } from './user';
+
+export const handlers = [...eventHandlers, ...resumeHandlers, ...userHandlers];
